requirements.txt
streamlit==1.10.0
requests==2.27.1
pytz
python-dotenv
Babel
